requirements.txt
wheel
tkinter
paho-mqtt
playsound==1.2.2